v0.1.0 (#60)
* Bump @testing-library/jest-dom from 4.2.4 to 5.11.4 in /frontend

Bumps [@testing-library/jest-dom](https://github.com/testing-library/jest-dom) from 4.2.4 to 5.11.4.
- [Release notes](https://github.com/testing-library/jest-dom/releases)
- [Changelog](https://github.com/testing-library/jest-dom/blob/master/CHANGELOG.md)
- [Commits](https://github.com/testing-library/jest-dom/compare/v4.2.4...v5.11.4)


* Add sleep for better build stability

* Fixing issues with domainDb

* Bump uuid from 8.3.0 to 8.3.1 in /api (#38)

Bumps [uuid](https://github.com/uuidjs/uuid) from 8.3.0 to 8.3.1.
- [Release notes](https://github.com/uuidjs/uuid/releases)
- [Changelog](https://github.com/uuidjs/uuid/blob/master/CHANGELOG.md)
- [Commits](https://github.com/uuidjs/uuid/compare/v8.3.0...v8.3.1)



* Bump @testing-library/react from 9.5.0 to 11.0.4 in /frontend (#58)

Bumps [@testing-library/react](https://github.com/testing-library/react-testing-library) from 9.5.0 to 11.0.4.
- [Release notes](https://github.com/testing-library/react-testing-library/releases)
- [Changelog](https://github.com/testing-library/react-testing-library/blob/master/CHANGELOG.md)
- [Commits](https://github.com/testing-library/react-testing-library/compare/v9.5.0...v11.0.4)



* Fix issues updating subdomain ip value on do

* only update if DigitalOcean is initialized

* Fixes [DOD-2]: Resolved issues with subdomains not updating the ip address

* Fixes [DOD-9]: Added subdomain validation

* Fixes [DOD-16]: Fixing issues with subdomains not being refreshed with DigitalOcean

* Fixes [DOD-8]: Added validation logic to settings routes

* Fixes [DOD-7]: Added validation logic to domain routes

* Move developement api to port 3081 while production still uses 3080

* Fix production ports

* Fixes [DOD-19]: Network Update Interval is now defined in minutes rather than milliseconds

* Resolved issues updating settings

* Fixes [DOD-18]: DigitalOcean data is now refreshed when the API token is supplied.

* Cleanup

* Fixes [DOD-17]: Subdomains now immediately update in the UI when set to active

* Resolve issue where development would kill any running production containers

* Bump version to v0.1.0

* Fix production container names

* Update azure-pipelines.yml for Azure Pipelines

+++ api/src/route/domain.js
@@ -25,15 +25,18 @@ router.get("/", (req, res) => {
  * Create a new domain record.
  */
 router.post("/", (req, res) => {
-  const { domain } = req.body;
-  // TODO: validate domain
+  const domain = req.body;
+  if (!validateDomain(domain)) {
+    res.status(500).json({ error: "malformed data" });
+    return;
+  }
 
-  db.find({ domain })
+  db.find({ name: domain.name })
     .then(d => {
       if (d)
         Promise.reject("The domain has already been created.")
     })
-    .then(() => db.insert({ hostname, domain, active: true }))
+    .then(() => db.insert(domain))
     .then(domain => {
       res.json(domain);
       logManager.addLog(`Registered new domain "${domain.domain}"`);
@@ -48,11 +51,14 @@ router.post("/", (req, res) => {
  */
 router.put("/:id", (req, res) => {
   const domain = req.body;
-  // TODO: validate domain
-  db.exists(req.param.id)
+  if (!validateDomain(domain)) {
+    res.status(500).json({ error: "no 'id' supplied in url" });
+    return;
+  }
+  db.exists(req.params.id)
     .then(exists => {
-      if (exists)
-        return Promise.reject(`The domain with id ${req.param.id} cannot be found in the database.`);
+      if (!exists)
+        return Promise.reject(`The domain with id ${req.paramsa.id} cannot be found in the database.`);
     })
     .then(() => db.update(domain))
     .then(domain => {
@@ -73,10 +79,24 @@ router.put("/:id", (req, res) => {
  */
 router.delete("/:id", (req, res) => {
   const { id } = req.params;
-  // TODO: id
+  if (!id) {
+    res.status(500).json({ error: "no 'id' supplied in url" });
+    return;
+  }
   db.delete(id)
     .then(domain => res.json(domain))
     .catch(err => res.status(500).json({ error: err }));
 });
+
+/**
+ * Validate a domain.
+ * @param {{ _id: string, name: string, ttl: number, zone_file: string, active: boolean, recordCreated: number, recordUpdated: number }} domain the domain.
+ */
+const validateDomain = (domain) => {
+  if (!domain._id || !domain.ttl || !domain.name || !domain.zone_file) {
+    return false;
+  }
+  return true
+};
 
 module.exports = router;
